+++ tests/integration_tests/test_rollback.py
@@ -137,6 +137,7 @@ def get_dipdup() -> DipDup:
     config = DipDupConfig.load([join(dirname(__file__), 'hic_et_nunc.yml')])
     config.database.path = ':memory:'  # type: ignore
     config.indexes['hen_mainnet'].last_level = 0  # type: ignore
+    config.advanced.early_realtime = True
     config.initialize()
     return DipDup(config)
 
@@ -145,34 +146,34 @@ class RollbackTest(IsolatedAsyncioTestCase):
     async def test_rollback_exact(self):
         with patch_dipdup(datasource_run_exact):
             dipdup = get_dipdup()
-            await dipdup.run(early_realtime=True)
+            await dipdup.run()
 
             assert dipdup._ctx.reindex.call_count == 0
 
     async def test_rollback_more(self):
         with patch_dipdup(datasource_run_more):
             dipdup = get_dipdup()
-            await dipdup.run(early_realtime=True)
+            await dipdup.run()
 
             assert dipdup._ctx.reindex.call_count == 0
 
     async def test_rollback_less(self):
         with patch_dipdup(datasource_run_less):
             dipdup = get_dipdup()
-            await dipdup.run(early_realtime=True)
+            await dipdup.run()
 
             assert dipdup._ctx.reindex.call_count == 1
 
     async def test_rollback_zero(self):
         with patch_dipdup(datasource_run_zero):
             dipdup = get_dipdup()
-            await dipdup.run(early_realtime=True)
+            await dipdup.run()
 
             assert dipdup._ctx.reindex.call_count == 0
 
     async def test_rollback_deep(self):
         with patch_dipdup(datasource_run_deep):
             dipdup = get_dipdup()
-            await dipdup.run(early_realtime=True)
+            await dipdup.run()
 
             assert dipdup._ctx.reindex.call_count == 1